refactor(cashier-v2): restructure payment agent modules folder
