+++ packages/cashier-v2/src/lib/PaymentAgent/PaymentAgentDeposit/__tests__/PaymentAgentDeposit.spec.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import PaymentAgentDeposit from '../PaymentAgentDeposit';
-import { usePaymentAgentContext } from '../../../provider';
+import { usePaymentAgentContext } from '../../provider';
 
 const mockUsePaymentAgentContext = usePaymentAgentContext as jest.MockedFunction<typeof usePaymentAgentContext>;
 
@@ -10,14 +10,14 @@ jest.mock('@deriv-com/ui', () => ({
     Loader: () => <div>Loader</div>,
 }));
 
-jest.mock('../../../provider', () => ({
+jest.mock('../../provider', () => ({
     PaymentAgentProvider: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
     usePaymentAgentContext: jest.fn(() => ({
         isPaymentAgentListLoading: false,
     })),
 }));
 
-jest.mock('../../../components', () => ({
+jest.mock('../../components', () => ({
     PaymentAgentList: () => <div>PaymentAgentList</div>,
     PaymentAgentSearchContainer: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
 }));
