+++ packages/cashier-v2/src/lib/PaymentAgent/PaymentAgentWithdrawal/components/PaymentAgentWithdrawalReceipt/PaymentAgentReceiptDetail/index.ts
