+++ packages/cashier-v2/src/lib/PaymentAgent/PaymentAgentWithdrawal/components/PaymentAgentWithdrawalReceipt/PaymentAgentWithdrawalReceipt.module.scss
