+++ packages/cashier-v2/src/lib/PaymentAgent/PaymentAgentWithdrawal/components/PaymentAgentWithdrawalReceipt/PaymentAgentWithdrawalReceipt.tsx
@@ -4,9 +4,9 @@ import { useHistory } from 'react-router-dom';
 import { StandalonePhoneRegularIcon } from '@deriv/quill-icons';
 import { Button, Text } from '@deriv-com/ui';
 import { FormatUtils } from '@deriv-com/utils';
-import EmailIcon from '../../../../../../assets/images/ic-email-outline-new.svg';
-import WebsiteIcon from '../../../../../../assets/images/ic-website.svg';
-import type { THooks } from '../../../../../../hooks/types';
+import EmailIcon from '../../../../../assets/images/ic-email-outline-new.svg';
+import WebsiteIcon from '../../../../../assets/images/ic-website.svg';
+import type { THooks } from '../../../../../hooks/types';
 import { usePaymentAgentWithdrawalContext } from '../../provider';
 import { PaymentAgentReceiptDetail } from './PaymentAgentReceiptDetail';
 import styles from './PaymentAgentWithdrawalReceipt.module.scss';
